refactor: change min sdk, refactor explorer to show full build
