+++ src/explorer/nodes.ts
@@ -1,11 +1,11 @@
 export * from './nodes/baseNode';
 export * from './nodes/blankNode';
 export * from './nodes/commandNode';
-export * from './nodes/deviceNode';
-export * from './nodes/distributeNode';
-export * from './nodes/osVerNode';
-export * from './nodes/platformNode';
-export * from './nodes/targetNode';
-export * from './nodes/updateNode';
-export * from './nodes/updatesNode';
+export * from './nodes/build/deviceNode';
+export * from './nodes/build/distributeNode';
+export * from './nodes/build/osVerNode';
+export * from './nodes/build/platformNode';
+export * from './nodes/build/targetNode';
+export * from './nodes/updates/updateNode';
+export * from './nodes/updates/updatesNode';
 export * from './nodes/urlNode';
